external_site/requirements.txt
Flask>=3.0
python-dotenv>=1.0
requests>=2.28
